+++ src/main/resources/application.properties
@@ -1,4 +1,7 @@
 spring.jpa.hibernate.ddl-auto=create-drop
-spring.datasource.url=jdbc:mysql://localhost:3306/people_pictures?useSSL=false
+spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
+spring.jpa.show-sql=false
+spring.jpa.database=mysql
+spring.datasource.url=jdbc:mysql://localhost:3306/people_pictures?useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username=people_pictures_user
 spring.datasource.password=xtzcnjiufk
